# docs/swagger.yaml
basePath: /api
definitions:
...
  controllers.PayReadingJSON:
    properties:
      account_id:
        type: string
    required:
    - account_id
    type: object
...
  models.Asset:
    properties:
      address:
        description: Нерухомість
        type: string
      area:
        type: number
      cadastral_num:
        type: string
      created_at:
        type: integer
      currency:
        type: string
      current_price:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      depreciation_type:
        description: Амортизація
        type: string
      documents:
        description: "\U0001F525 ДОДАНО: Документи (PDF та інші)"
        items:
          $ref: '#/definitions/models.AssetDocument'
        type: array
      estimated_life:
        type: integer
      family_id:
        type: string
      id:
        type: string
      initial_mileage:
        type: integer
      initial_value:
        type: integer
      insurance_expiry:
        type: integer
      is_sold:
        description: Продаж
        type: boolean
      is_synced:
        type: boolean
      last_service_date:
        type: integer
      mileage:
        type: integer
      name:
        type: string
      note:
        type: string
      photo:
        description: Медіа
        type: string
      photos:
        items:
          $ref: '#/definitions/models.AssetPhoto'
        type: array
      price:
        description: Фінанси
        type: integer
      purchase_date:
        type: integer
      serial_number:
        description: Деталі
        type: string
      sold_date:
        type: integer
      sold_price:
        type: integer
      type:
        description: '''car'', ''real_estate'', ''appliance'', ''furniture'''
        type: string
      updated_at:
        type: integer
      user_id:
        type: string
      vin_code:
        description: Автоспецифіка
        type: string
      warranty_end:
        type: integer
    type: object
  models.AssetDocument:
    properties:
      asset_id:
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      file_type:
        description: Розширення (напр. "pdf")
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      name:
        description: Оригінальна назва файлу (напр. "чек.pdf")
        type: string
      path:
        description: Шлях на сервері
        type: string
      size:
        description: Розмір в байтах (опціонально, але корисно)
        type: integer
      updated_at:
        type: integer
    type: object
  models.AssetPhoto:
    properties:
      asset_id:
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      id:
        type: string
      is_synced:
        type: boolean
      path:
        type: string
      updated_at:
        type: integer
    type: object
...
  models.Counterparty:
    properties:
      balances:
        description: Баланси (борги)
        items:
          $ref: '#/definitions/models.CounterpartyBalance'
        type: array
      category:
        $ref: '#/definitions/models.CounterpartyCategory'
      category_id:
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        type: string
      icon:
        type: string
      id:
        type: string
      is_group:
        type: boolean
      is_synced:
        type: boolean
      logo:
        type: string
      name:
        type: string
      parent_id:
        type: string
      type:
        description: '''shop'', ''person'', ''other'''
        type: string
      updated_at:
        type: integer
    type: object
  models.CounterpartyBalance:
    properties:
      balance:
        description: + нам винні, - ми винні
        type: integer
      counterparty_id:
        type: string
      currency:
        type: string
    type: object
  models.CounterpartyCategory:
    properties:
      color:
        type: string
      created_at:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        type: string
      icon:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      name:
        type: string
      type:
        type: string
      updated_at:
        type: integer
    type: object
  models.CreateAssetOnFlyInput:
    properties:
      address:
        type: string
      area:
        type: number
      insurance_expiry:
        type: integer
      last_service_date:
        type: integer
      mileage:
        type: integer
      name:
        type: string
      note:
        type: string
      price:
        type: integer
      purchase_date:
        type: integer
      serial_number:
        type: string
      type:
        type: string
      vin_code:
        type: string
      warranty_end:
        type: integer
    type: object
...
  models.UtilityMeter:
    properties:
      asset:
        $ref: '#/definitions/models.Asset'
      asset_id:
        description: Прив'язки
        type: string
      counterparty:
        $ref: '#/definitions/models.Counterparty'
      counterparty_id:
        type: string
      created_at:
        type: integer
      currency:
        type: string
      deleted_at:
        description: Індекс для soft delete
        type: integer
      family_id:
        type: string
      id:
        type: string
      is_synced:
        type: boolean
      last_reading_date:
        description: Кеш останніх показників
        type: integer
      last_reading_value:
        type: number
      name:
        type: string
      new_asset:
        allOf:
        - $ref: '#/definitions/models.CreateAssetOnFlyInput'
        description: Хак для створення Asset "на льоту"
      personal_account:
        type: string
      tariff:
        type: number
      type:
        description: electricity, water, gas, internet...
        type: string
      unit:
        description: kW, m3, Gcal
        type: string
      updated_at:
        type: integer
    type: object
  models.UtilityReading:
    properties:
      calculated_cost:
        type: integer
      created_at:
        type: integer
      date:
        type: integer
      deleted_at:
        description: Індекс для soft delete
        type: integer
      diff:
        type: number
      id:
        type: string
      is_paid:
        type: boolean
      is_synced:
        type: boolean
      meter:
        $ref: '#/definitions/models.UtilityMeter'
      meter_id:
        type: string
      payment_transaction_id:
        description: 'Якщо у тебе логіка розділяє оплату і нарахування, залишаємо
          і це:'
        type: string
      tariff_at_date:
        type: number
      transaction_id:
        description: "--- \U0001F447 ПОВЕРНУВ TransactionID ---\nЦе поле використовується
          для зв'язку з транзакцією оплати"
        type: string
      updated_at:
        type: integer
      value:
        type: number
    type: object
  models.UtilityStatGlobalDTO:
    properties:
      data:
        additionalProperties:
          format: float64
          type: number
        type: object
      month:
        type: string
    type: object
  models.UtilityStatMeterDTO:
    properties:
      avg_tariff:
        description: Середній тариф за місяць
        type: number
      month:
        type: string
      total_consumption:
        description: Спожито одиниць (кВт, м3)
        type: number
      total_cost:
        description: Гроші (копійки)
        type: number
    type: object
...
info:
  contact:
    name: Vlad Hrytsaiuk
  description: API for WeGaS Finance application.
  title: WeGaS Finance API
  version: "1.0"
paths:
...
  /utility/meters:
    get:
      consumes:
      - application/json
      description: Returns a list of all utility meters for the current family.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UtilityMeter'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get all utility meters
      tags:
      - Utility
    post:
      consumes:
      - application/json
      description: Creates a new utility meter (electricity, water, etc.).
      parameters:
      - description: Meter details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UtilityMeter'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new utility meter
      tags:
      - Utility
  /utility/meters/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a utility meter by its ID.
      parameters:
      - description: Meter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete status
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete utility meter
      tags:
      - Utility
    get:
      consumes:
      - application/json
      description: Returns a single utility meter by its ID.
      parameters:
      - description: Meter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UtilityMeter'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Meter not found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get utility meter by ID
      tags:
      - Utility
    put:
      consumes:
      - application/json
      description: Updates an existing utility meter by its ID.
      parameters:
      - description: Meter ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated meter details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UtilityMeter'
      produces:
      - application/json
      responses:
        "200":
          description: Update status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update utility meter
      tags:
      - Utility
  /utility/readings:
    get:
      consumes:
      - application/json
      description: Returns a list of utility readings, optionally filtered by meter
        ID.
      parameters:
      - description: Filter by meter ID
        in: query
        name: meter_id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UtilityReading'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get utility readings
      tags:
      - Utility
    post:
      consumes:
      - application/json
      description: Records a new reading for a utility meter.
      parameters:
      - description: Reading details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UtilityReading'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Invalid input or reading sequence
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Create a new utility reading
      tags:
      - Utility
  /utility/readings/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a utility reading by its ID.
      parameters:
      - description: Reading ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Delete status
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete utility reading
      tags:
      - Utility
    put:
      consumes:
      - application/json
      description: Updates an existing utility reading by its ID.
      parameters:
      - description: Reading ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated reading details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.UtilityReading'
      produces:
      - application/json
      responses:
        "200":
          description: Update status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Update utility reading
      tags:
      - Utility
  /utility/readings/{id}/pay:
    post:
      consumes:
      - application/json
      description: Processes a payment for a specific utility reading using a financial
        account.
      parameters:
      - description: Reading ID
        in: path
        name: id
        required: true
        type: string
      - description: Payment details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/controllers.PayReadingJSON'
      produces:
      - application/json
      responses:
        "200":
          description: Payment status
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Pay for a utility reading
      tags:
      - Utility
  /utility/stats/global:
    get:
      consumes:
      - application/json
      description: Returns aggregated consumption and cost statistics for all utility
        meters.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UtilityStatGlobalDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get global utility statistics
      tags:
      - Utility
  /utility/stats/meter/{id}:
    get:
      consumes:
      - application/json
      description: Returns consumption and cost trend statistics for a single utility
        meter.
      parameters:
      - description: Meter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.UtilityStatMeterDTO'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get statistics for a specific utility meter
      tags:
      - Utility
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"